+++ resources/views/gallery/album-show.blade.php
@@ -31,46 +31,7 @@ $stats = array_filter([
     </x-slot>
   </x-sections.header>
 
-  <x-title>
-    <h1>Album {{ $album->name }}</h1>
-
-    <x-slot name="subtitle">
-      @if ($album->photos->count() > 0)
-        {{ $album->photos_count ?? $album->photos->count() }} foto's, genomen vanaf {{
-        $album->photos->min('taken_at')?->isoFormat('DD MMMM YYYY') }}
-      @else
-        Leeg album
-      @endif
-    </x-slot>
-</x-title>
-
-<div class="bg-gray-100 mb-4">
-  <div class="container py-4 grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-4 gap-4">
-    @if ($album->activity)
-        <div class="flex items-center">
-          <x-icon icon="calendar-alt" class="h-4 mr-4" />
-          <span>Hoort bij <a href="{{ route('activity.show', $album->activity) }}">{{ $album->activity->name }}</a></span>
-        </div>
-    @elseif ($album->user)
-      <div class="flex items-center">
-        <x-icon icon="solid/user" class="h-4 mr-4" />
-        <span>Gemaakt door <strong>{{ $album->user->name }}</strong></span>
-      </div>
-    @endif
-
-    <div class="flex items-center">
-      <x-icon icon="solid/lightbulb" class="h-4 mr-4" />
-      <span>Aangemaakt op <strong>{{ $album->created_at->isoFormat('d MMM YYYY') }}</strong></span>
-    </div>
-
-    <div class="flex items-center">
-      <x-icon icon="solid/pencil-alt" class="h-4 mr-4" />
-      <span>Laatst bewerkt op <strong>{{ $album->updated_at->isoFormat('d MMM YYYY') }}</strong></span>
-    </div>
-  </div>
-</div>
-
-<div class="container">
+<x-container>
   <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-8">
     @forelse ($album->photos as $photo)
       <x-gallery.photo-tile :photo="$photo" />
@@ -82,5 +43,5 @@ $stats = array_filter([
       </div>
     @endforelse
   </div>
-</div>
+</x-container>
 </x-page>
